+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+* Add `label.overflow` option to truncate labels that are too wide for their blocks with an ellipsis (#241)
 * Add `label.verticalAlign` option to align in-block labels to the `top`, `middle`, or `bottom` of their blocks (#96)
 * Add `block.gap` option to leave space between blocks (#88)